refactor: move dock page API to workflow module
