+++ ui/dockwidget/wizard_composition.py
@@ -47,7 +47,7 @@ from .step_page import (
     apply_workflow_step_page_statuses,
     install_workflow_step_pages,
 )
-from .wizard_page import WorkflowPage, install_workflow_pages
+from .workflow_page import WorkflowPage, install_workflow_pages
 from .wizard_shell import WorkflowShell
 from .workflow_shell_presenter import WorkflowShellPresenter
 from .workflow_page_state import (
